Import templates from parent module (#408)
* add __all__ variables to submodules

* Shorten/format some docstrings.

* accidentally have changed from other branch

* sorted submodules alphabetically in autosummary

* outsourced interferometer to new module

* add subroutines to docs intro

* deleted unnecessary imports that pycharm did not highlight

* deleted unnecessary imports that pycharm did not highlight

* implement Nathan's review suggestions
